refactor(runtime): #5590 — split thenable assimilation out of promise/combinators.rs
The #5590 fix pushed combinators.rs over the 2000-line per-file size gate
(scripts/check_file_size.sh). Extract the cohesive resolve-with-thenable cluster
into a new sibling module promise/assimilate.rs (the sanctioned remedy — not an
allowlist entry):

  enum OwnThen, promise_own_then, get_then_action,
  get_array_prototype_then_action, is_native_array_value, enqueue_thenable_job,
  thenable_job_take_guard / _resolve_fn / _reject_fn, promise_resolve_thenable_job,
  promise_resolve_assimilating, assimilate_via_then_property
  (+ the non-callable-own-then unit test).

Wiring: `pub mod assimilate;` in promise/mod.rs; combinators.rs imports what it
still calls from the new module (and exposes `ensure_native_resolving_arity_
registered` as pub(super)); the external callers in async_step.rs and
microtasks.rs point at promise::assimilate::promise_resolve_assimilating.

Pure code motion — no behavior change. combinators.rs 2053 → 1702 lines;
check_file_size.sh OK; cargo fmt clean. perry-runtime promise tests pass; the
built-ins/Promise slice is unchanged at 532 pass / 42 fail (identical failure
set).
